Backend/requirements.txt
requests
pymongo
bcrypt
python-dotenv
argparse
langchain
langchain_huggingface
huggingface_hub
langchain_community
streamlit
authlib
plotly
Authlib
streamlit
bcrypt
python-dotenv
matplotlib
psutil